requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
django-environ==0.11.2
django-graphql-jwt==0.4.0
django-guardian==2.4.0
django-pgcrypto==2.0.0
django-pgcrypto-fields==2.6.0
django-simple-history==3.5.0
graphene==3.3
graphene-django==3.2.0
graphene-django-cud==0.11.1
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
install==1.3.5
Jinja2==3.1.3
MarkupSafe==2.1.5
promise==2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
